service/requirements.txt
telethon
requests
beautifulsoup4
lxml
pyyaml
schedule